Add server demo executable and enhance curve and PRG implementations
- Introduced a new server demo executable for the poker game simulation.
- Updated curve initialization to use NIST P-256 parameters.
- Modified Scalar creation to accept unsigned long long for better range.
- Enhanced PRG class for improved randomness handling across platforms.

+++ src/curve.h
@@ -24,7 +24,7 @@ class Scalar {
   friend class Point;
 
   static Scalar CreateRandom();
-  static Scalar CreateFromInt(unsigned int v);
+  static Scalar CreateFromInt(unsigned long long v);
   static Scalar Read(const uint8_t* bytes);
 
   static constexpr std::size_t ByteSize() { return 32; };
